Split out Hardware from the core. This will make it easier to add new definitions.
